refactor: dto와 도메인이 변경됨에 따라 Mapper가 변경됨
