+++ android/app/src/main/java/com/now/naaga/data/mapper/AdventureMapper.kt
@@ -2,23 +2,13 @@ package com.now.naaga.data.mapper
 
 import com.now.domain.model.Adventure
 import com.now.domain.model.AdventureStatus
-import com.now.domain.model.Game
 import com.now.domain.model.RemainingTryCount
 import com.now.naaga.data.remote.dto.AdventureDto
 import com.now.naaga.data.remote.dto.EndedAdventureDto
-import com.now.naaga.data.remote.dto.third.GameDto
 import java.time.LocalDateTime
 
 fun AdventureDto.toDomain(): Adventure {
     return Adventure(
-        id = id,
-        destination = destinationDto.toDomain(),
-        adventureStatus = AdventureStatus.getStatus(adventureStatus),
-    )
-}
-
-fun GameDto.toDomain(): Game {
-    return Game(
         id = id,
         createdAt = LocalDateTime.parse(startTime),
         startCoordinate = startCoordinate.toDomain(),
